feat(handlers): POST /posts

+++ internal/models/post.go
@@ -6,7 +6,7 @@ type Post struct {
 	ID      *int   `json:"id"`
 	Title   string `json:"title" binding:"required"`
 	Content string `json:"content" binding:"required"`
-	UserID  uint64 `json:"user_id" binding:"required"`
+	UserID  int    `json:"user_id" binding:"required"`
 }
 
 func (p Post) ToEnt() *ent.Post {
